Converted App.js to Typescript
